Rename 16next.cpp/h to main.cpp/h
