+++ CMakeLists.txt
@@ -19,7 +19,7 @@ set(CMAKE_CXX_STANDARD 17)
 pico_sdk_init()
 
 add_executable(${target_proj}
-  16next.cpp
+  main.cpp
 )
 
 target_sources(${target_proj}
